GPII-3138:  Script to update snapsets.
Added more logging to see why the call to doBatchDelete() happens
before doBatchDelete() is called.  The bug is that only the snapset
Prefs Safes are deleted.  Their GPII Keys are added to the 'docsToRemove'
array after that deletion.  I'm not sure why, but suspect it has to
do with some kind of asynchrony in the http responses. Also, I have not
found the condition to test when all the GPII Key records have been
added.