GPII-3138:  Update snapsets in the database
Modified the way that universal is set up within the Docker image
in order that the latest version of universal is used.  This is
done by cloning from github when the image is run, not when
the image is built.  Also, this means that the preferences are
converted to Prefs Safes and GPII Keys at "run-time", not
"build-time".